refactor: move internal packages to server
