+++ testscript/script_test.go
@@ -103,6 +103,7 @@ func TestScript(t *testing.T) {
 
 			ctx := config.WithContext(context.Background(), &cfg)
 
+			// TODO: test postgres
 			db, err := db.Open(ctx, cfg.DB.Driver, cfg.DB.DataSource)
 			if err != nil {
 				return fmt.Errorf("open database: %w", err)
